# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ElevateKe Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  phone: string;
  password: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  balance: number;
  totalEarned: number;
  totalDeposited: number;
  vipLevel: string;
  referralCode: string;
  isAdmin: boolean;
  isSuspended: boolean;
  /** @nullable */
  loginBonusClaimedAt?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  mpesaPhone?: string | null;
  balance: number;
  totalEarned: number;
  totalDeposited: number;
  vipLevel: string;
  referralCode: string;
  language: string;
  createdAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  phone?: string;
  mpesaPhone?: string;
  language?: string;
}

export interface ReferralEntry {
  id: number;
  name: string;
  level: number;
  bonusAmount: number;
  joinedAt: string;
}

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
  level1Count: number;
  level2Count: number;
  totalReferralEarnings: number;
  referrals: ReferralEntry[];
}

export interface LoginHistoryEntry {
  id: number;
  /** @nullable */
  ip?: string | null;
  createdAt: string;
}

export interface BonusResult {
  amount: number;
  newBalance: number;
  message: string;
}

export interface Trade {
  id: number;
  amount: number;
  multiplier: number;
  durationMins: number;
  direction: string;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  profitLoss?: number | null;
  status: string;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface DashboardSummary {
  balance: number;
  totalEarned: number;
  totalDeposited: number;
  pendingWithdrawals: number;
  activeDeposits: number;
  /** @nullable */
  nextEarningAt?: string | null;
  todayEarned: number;
  vipLevel: string;
  unreadMessages: number;
  loginBonusAvailable: boolean;
  activeTrade?: Trade;
}

export interface DepositPlan {
  id: number;
  name: string;
  minAmount: number;
  /** @nullable */
  maxAmount?: number | null;
  dailyRate: number;
  dailyEarning: number;
  durationDays: number;
  bonusPercent: number;
  isActive: boolean;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CreatePlanBody {
  name: string;
  minAmount: number;
  /** @nullable */
  maxAmount?: number | null;
  dailyRate: number;
  durationDays: number;
  bonusPercent: number;
  isActive: boolean;
  /** @nullable */
  description?: string | null;
}

export interface Deposit {
  id: number;
  userId: number;
  planId: number;
  planName: string;
  amount: number;
  bonusAmount: number;
  dailyEarning: number;
  status: string;
  autoRenew: boolean;
  /** @nullable */
  startsAt?: string | null;
  /** @nullable */
  endsAt?: string | null;
  /** @nullable */
  lastEarningAt?: string | null;
  createdAt: string;
}

export interface CreateDepositBody {
  planId: number;
  amount: number;
  autoRenew?: boolean;
  phone: string;
}

export interface DepositWithPaystack {
  deposit: Deposit;
  paystackAuthUrl: string;
  reference: string;
}

export interface VerifyDepositBody {
  reference: string;
}

export interface Earning {
  id: number;
  amount: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface ReinvestBody {
  amount: number;
}

export interface Task {
  id: number;
  title: string;
  description: string;
  reward: number;
  /** @nullable */
  link?: string | null;
  isActive: boolean;
  isCompleted: boolean;
  createdAt: string;
}

export interface CreateTaskBody {
  title: string;
  description: string;
  reward: number;
  /** @nullable */
  link?: string | null;
  isActive: boolean;
}

export interface Withdrawal {
  id: number;
  amount: number;
  phone: string;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  requestedAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface CreateWithdrawalBody {
  amount: number;
  phone: string;
  pin: string;
}

export interface TradeSettings {
  direction: string;
  updatedAt: string;
}

export interface ChartPoint {
  time: string;
  value: number;
}

export interface PlaceTradeBody {
  amount: number;
  multiplier: number;
  durationMins: number;
}

export interface CashoutTradeBody {
  tradeId: number;
}

export interface TradeResult {
  trade: Trade;
  profitLoss: number;
  newBalance: number;
  message: string;
}

export interface UpdateTradeDirectionBody {
  direction: string;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateAnnouncementBody {
  title: string;
  content: string;
  isActive: boolean;
}

export interface InboxMessage {
  id: number;
  title: string;
  content: string;
  isRead: boolean;
  createdAt: string;
}

export interface SendInboxMessageBody {
  /** @nullable */
  userId?: number | null;
  title: string;
  content: string;
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  suspendedUsers: number;
  totalDeposits: number;
  totalWithdrawals: number;
  totalEarnings: number;
  pendingWithdrawals: number;
  pendingWithdrawalsCount: number;
  totalReferrals: number;
  todayDeposits: number;
  todayWithdrawals: number;
  todayNewUsers: number;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  phone: string;
  balance: number;
  totalEarned: number;
  totalDeposited: number;
  vipLevel: string;
  referralCode: string;
  referralCount: number;
  isSuspended: boolean;
  createdAt: string;
}

export interface AdminUserDetail {
  user: AdminUser;
  deposits: Deposit[];
  withdrawals: Withdrawal[];
  earnings: Earning[];
  referrals: ReferralEntry[];
}

export interface AdminWithdrawal {
  id: number;
  userId: number;
  userName: string;
  userPhone: string;
  amount: number;
  phone: string;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  requestedAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface RejectWithdrawalBody {
  reason: string;
}

export interface ReportData {
  users: AdminUser[];
  totalStats: AdminStats;
}