# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ElevateKe Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admin-login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /deposits:
    get:
      operationId: getDeposits
      tags: [deposits]
      summary: Get user deposits
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Create a deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepositBody"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositWithPaystack"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deposits/verify:
    post:
      operationId: verifyDeposit
      tags: [deposits]
      summary: Verify Paystack payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyDepositBody"
      responses:
        "200":
          description: Deposit verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
        "400":
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /earnings/claim:
    post:
      operationId: claimDailyEarnings
      tags: [earnings]
      summary: Claim daily earnings
      responses:
        "200":
          description: Earnings claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusResult"
        "400":
          description: Already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /earnings/reinvest:
    post:
      operationId: reinvestEarnings
      tags: [earnings]
      summary: Reinvest earnings into balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReinvestBody"
      responses:
        "200":
          description: Reinvested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /inbox/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [inbox]
      summary: Mark message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
      required: [name, email, phone, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        balance:
          type: number
        totalEarned:
          type: number
        totalDeposited:
          type: number
        vipLevel:
          type: string
        referralCode:
          type: string
        isAdmin:
          type: boolean
        isSuspended:
          type: boolean
        loginBonusClaimedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, phone, balance, totalEarned, totalDeposited, vipLevel, referralCode, isAdmin, isSuspended, createdAt]
...
    BonusResult:
      type: object
      properties:
        amount:
          type: number
        newBalance:
          type: number
        message:
          type: string
      required: [amount, newBalance, message]
...
    Deposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        amount:
          type: number
        bonusAmount:
          type: number
        dailyEarning:
          type: number
        status:
          type: string
        autoRenew:
          type: boolean
        startsAt:
          type: ["string", "null"]
        endsAt:
          type: ["string", "null"]
        lastEarningAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, planId, planName, amount, bonusAmount, dailyEarning, status, autoRenew, createdAt]
    CreateDepositBody:
      type: object
      properties:
        planId:
          type: integer
        amount:
          type: number
        autoRenew:
          type: boolean
        phone:
          type: string
      required: [planId, amount, phone]
    DepositWithPaystack:
      type: object
      properties:
        deposit:
          $ref: "#/components/schemas/Deposit"
        paystackAuthUrl:
          type: string
        reference:
          type: string
      required: [deposit, paystackAuthUrl, reference]
    VerifyDepositBody:
      type: object
      properties:
        reference:
          type: string
      required: [reference]
...
    ReinvestBody:
      type: object
      properties:
        amount:
          type: number
      required: [amount]
...